Tehtävät 4.23

+++ osa4/blog/controllers/blogs.js
@@ -1,6 +1,9 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const { blogsInDb } = require('../tests/test_helper/test_helper')
+const req = require('express/lib/request')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
@@ -9,9 +12,16 @@ blogsRouter.get('/', async (request, response) => {
   })
   
   blogsRouter.post('/', async (request, response) => {
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' }).end()
+    }
     const body = request.body
+    const userId = request.userId
+    
+    const user = await User.findById(userId)
 
-    const user = await User.findById(body.userId)
     const blog = new Blog({
       title: body.title,
       author: body.author,
@@ -26,13 +36,36 @@ blogsRouter.get('/', async (request, response) => {
   })
 
   blogsRouter.delete('/:id', async (request, response) => {
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' }).end()
+    }
+
+    const userId = request.userId
+    const blog = await Blog.findById(request.params.id)
+    if (!blog){
+      response
+      .status(404).end()
+    }
+    if ( blog.user.toString() !== userId.toString() ){
+      response.status(401).end()
+    }
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
   })
 
   blogsRouter.put('/:id', async (request, response) => {
     const body = request.body
-  
+    const userId = request.userId
+    const blog = await Blog.findById(request.params.id)
+    if (!blog){
+      response
+      .status(404).end()
+    }
+    if ( blog.user.toString() !== userId.toString() ){
+      response.status(401).end()
+    }
     const updatedBlog = await Blog.findByIdAndUpdate(
       request.params.id,
       {
